+++ src/views/stage/create/api/postFastStage.ts
@@ -3,7 +3,7 @@ import { StageData } from '@/shared/types/stage/create';
 
 export const postFastStage = async (data: StageData) => {
   try {
-    await axios.post('/api/stage/fast', data);
+    await axios.post('/api/server/stage/fast', data);
     return true;
   } catch (error) {
     if (axios.isAxiosError(error) && error.response) {
